Car's method sorted
Dodanie klas:
-CarsByDateService
-CarsByMileageService
-CarsByPriceService
-CarsBySpecificationService
przeniesienie odpowiednich metod do tych klas, wstrzyknięcie do tych klas Beana "carService" oraz zmiany klas wywołujących te metody w klasie kontrolera
