+++ src/main/java/com/example/rentalcars/MenagementServices/CarsBySpecificationService.java
@@ -1,0 +1,57 @@
+package com.example.rentalcars.MenagementServices;
+
+import com.example.rentalcars.enums.BodyType;
+import com.example.rentalcars.enums.FuelType;
+import com.example.rentalcars.enums.GearboxType;
+import com.example.rentalcars.model.CarModel;
+import com.example.rentalcars.repository.CarRepository;
+import com.example.rentalcars.service.CarService;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+
+import java.util.List;
+
+@Service
+@RequiredArgsConstructor
+public class CarsBySpecificationService {
+
+    private final CarRepository carRepository;
+
+    private final CarService carService;
+
+    public List<CarModel> getCarsByGearbox(GearboxType gearbox) {
+        return carService.getCarList().stream()
+                .filter(car -> car.getGearbox() != null)
+                .filter(car -> car.getGearbox().equals(gearbox))
+                .toList();
+    }
+
+    public List<CarModel> getCarsByMark(String mark) {
+        return carService.getCarList().stream()
+                .filter(car -> car.getMark() != null)
+                .filter(car -> car.getMark().equals(mark))
+                .toList();
+    }
+
+    public List<CarModel> getCarsByBodyType(BodyType bodyType) {
+        return carService.getCarList().stream()
+                .filter(car -> car.getBody() != null)
+                .filter(car -> car.getBody().equals(bodyType))
+                .toList();
+    }
+
+    public List<CarModel> getCarsByColor(String color) {
+        return carService.getCarList().stream()
+                .filter(car -> car.getColor() != null)
+                .filter(car -> car.getColor().equals(color))
+                .toList();
+    }
+
+    public List<CarModel> getCarsByFuelType(FuelType fuelType) {
+        return carService.getCarList().stream()
+                .filter(car -> car.getFuelType() != null)
+                .filter(car -> car.getFuelType().equals(fuelType))
+                .toList();
+    }
+
+}
