+++ src/main/java/com/example/rentalcars/service/CarService.java
@@ -58,109 +58,7 @@ public class CarService {
         }
     }
 
-    public List<CarModel> getCarsByGearbox(GearboxType gearbox) {
-        return getCarList().stream()
-                .filter(car -> car.getGearbox() != null)
-                .filter(car -> car.getGearbox().equals(gearbox))
-                .toList();
-    }
 
-    public List<CarModel> getCarsByMark(String mark) {
-        return getCarList().stream()
-                .filter(car -> car.getMark() != null)
-                .filter(car -> car.getMark().equals(mark))
-                .toList();
-    }
-
-    public List<CarModel> getCarsByFuelType(FuelType fuelType) {
-        return getCarList().stream()
-                .filter(car -> car.getFuelType() != null)
-                .filter(car -> car.getFuelType().equals(fuelType))
-                .toList();
-    }
-
-    public List<CarModel> getCarsByPriceAscending() {
-        return getCarList().stream().sorted(Comparator.comparing(CarModel::getPrice)).toList();
-    }
-
-
-    public List<CarModel> getCarsByPriceDescending() {
-        return getCarList().stream().sorted(Comparator.comparing(CarModel::getPrice).reversed()).toList();
-    }
-
-    public List<CarModel> getCarsByPriceRange(BigDecimal priceMin, BigDecimal priceMax) {
-        return getCarList().stream().filter(car -> car.getPrice().compareTo(priceMin) >= 0 && car.getPrice().compareTo(priceMax) <= 0).toList();
-    }
-
-    public List<CarModel> getCarsByBodyType(BodyType bodyType) {
-        return getCarList().stream()
-                .filter(car -> car.getBody() != null)
-                .filter(car -> car.getBody().equals(bodyType))
-                .toList();
-    }
-
-    public List<CarModel> getCarsCheaperThan(BigDecimal upperPriceBorder) {
-        return getCarList().stream()
-                .filter(car -> car.getPrice() != null)
-                .filter(car -> car.getPrice().compareTo(upperPriceBorder) < 0)
-                .collect(toList());
-    }
-
-    public List<CarModel> getCarsMoreExpensiveThan(BigDecimal lowerPriceBorder) {
-        return getCarList().stream()
-                .filter(car -> car.getPrice() != null)
-                .filter(car -> car.getPrice().compareTo(lowerPriceBorder) >= 0)
-                .collect(toList());
-    }
-
-    public List<CarModel> getCarsWithHigherMileageThan(Integer mileage) {
-        return getCarList().stream()
-                .filter(car -> car.getMileage() != null)
-                .filter(car -> car.getMileage().compareTo(mileage) >= 0)
-                .toList();
-    }
-
-    public List<CarModel> getCarsWithLowerMileageThan(Integer mileage) {
-        return getCarList().stream()
-                .filter(car -> car.getMileage() != null)
-                .filter(car -> car.getMileage().compareTo(mileage) <= 0)
-                .toList();
-    }
-
-    public List<CarModel> getCarsWithinGivenMileage(Integer maxMileage, Integer minMileage) {
-        return getCarList().stream()
-                .filter(car -> car.getMileage() != null)
-                .filter(car -> car.getMileage().compareTo(minMileage) <= 0 && car.getMileage().compareTo(maxMileage) >= 0)
-                .toList();
-    }
-
-    public List<CarModel> getCarsByColor(String color) {
-        return getCarList().stream()
-                .filter(car -> car.getColor() != null)
-                .filter(car -> car.getColor().equals(color))
-                .toList();
-    }
-
-    public List<CarModel> getCarsByProductionDate(Integer productionDate) {
-        return getCarList().stream()
-                .filter(car -> car.getProductionDate() != null)
-                .filter(car -> car.getProductionDate().equals(productionDate))
-                .toList();
-    }
-
-    public List<CarModel> getCarsNewerThan(Integer productionDate) {
-        return getCarList().stream()
-                .filter(car -> car.getProductionDate() != null)
-                .filter(car -> car.getProductionDate().compareTo(productionDate) >= 0)
-                .toList();
-    }
-
-    public List<CarModel> getCarsOlderThan(Integer productionDate) {
-        return getCarList().stream()
-                .filter(car -> car.getProductionDate() != null)
-                .filter(car -> car.getProductionDate().compareTo(productionDate) <= 0)
-                .toList();
-    }
 
 
 }
